resume-counter/updateVisitorCountPython/requirements.txt
azure-functions==1.18.0
pytest

